CanvasGraphics: restore old code for SpreadMethod.PAD because the change in #2658 broke certain gradients
Some gradients were not rendering correctly in Chrome or Safari (but worked in Firefox, interestingly) with OpenFL 9.4, but worked fine in OpenFL 9.3.

Keeping the new SpreadMethod implementations the same, though. They might have a similar issue, but I'm not familiar enough with how they work to test to see if it reproduces.

Relevant community thread where issue was reported: https://community.openfl.org/t/issues-after-upgrading-to-openfl-9-4-0-lime-8-2-0/14026