CairoGraphics, CanvasGraphics: fix failed hit testing when graphics contains a mix of line styles
In playCommands(), it calls closePath() when LINE_STYLE is encountered, but we weren't doing a similar endStroke() call in hitTest() when LINE_STYLE is encountered. This meant that some of the strokes weren't getting checked for hit testing at all.

Now, when encountering LINE_STYLE in hitTest(), we call endStroke() and immediately check if the point is in the current stroke.

Context: https://community.openfl.org/t/line-thickness-and-hit-testing-pixel-level-behavior/14048